adds status components
