--- src/TaskComponents/CompletedTasks.js
@@ -1,16 +1,0 @@
-// import TaskItem from "./TaskItem";
-
-// function CompletedTasks({ tasks, onDoneTask, onDeleteTask }) {
-//   return (
-//     <ul className="completed-tasks">
-//       <h3>Completed</h3>
-//       <TaskItem
-//         tasks={tasks.filter((task) => task.status === "Completed")}
-//         onDone={onDoneTask}
-//         onDelete={onDeleteTask}
-//       />
-//     </ul>
-//   );
-// }
-
-// export default CompletedTasks;
