+++ src/TaskComponents/NavBar.css
@@ -5,9 +5,6 @@
     background-color: rgb(21, 123, 113);
     display: flex;
     justify-content: space-between;
-
-   
-
 }
 
 .nav-right{
